Documentation and general commenting done (see body)
As I was not well acquainted with the codebase, I needed some time to
familiarize myself with the constructs present, for I do not want to
violate the established paradigm of your API.

The controllers have Javadocs and general comments, and I have created a
helper method called debugLog() in each controller -- its a shorthand
System.out.println() that also shows the classname in the console, along
with a parameterized message. This can prove useful for testing to see
if a method call was made/and or verifying the count of a variable, etc.

There are a few classes that do not have documentation yet, but that can
be handled shortly.

More regular commits to come. Will next session begin by isolating the
Stock user and the creation of the image data in a separate method so
that I can resume work on the thumbnail display.